pyproject.toml
[project]
name = "mindstorms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "openai>=1.46.0",
    "python-dotenv>=1.0.1",
    "python-ev3dev2>=2.1.0.post1",
]
